farmerai_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.41.0
tensorflow-cpu==2.21.0
numpy==2.1.3
joblib==1.4.2
Pillow==11.1.0
python-dotenv==1.1.0
pandas==2.2.3
scikit-learn==1.6.1
httpx==0.28.1
python-multipart==0.0.22